更改int为long，方便greendao存储

+++ app/src/main/java/com/gaos/book/api/ApiFactory.java
@@ -34,7 +34,7 @@ public class ApiFactory {
         return Api.getInstance().service.getBookList().compose(RxSchedulers.observableIO2Main());
 
     }
-    public static Observable<BaseBean<List<CatalogInfo>>> getCatalogList(int book_id) {
+    public static Observable<BaseBean<List<CatalogInfo>>> getCatalogList(long book_id) {
         return Api.getInstance().service.getCatalogList(book_id).compose(RxSchedulers.observableIO2Main());
 
     }
